% Auto-generated: do not edit by hand
\name{''CardDropdown}

\alias{''CardDropdown}

\title{CardDropdown component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''CardDropdown(children=NULL, id=NULL, label=NULL, value=NULL, classes=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children components displayed inside the grid.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}

\item{classes}{Named list. The classes displayed in the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

