% Auto-generated: do not edit by hand
\name{''SelectFilter}

\alias{''SelectFilter}

\title{SelectFilter component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''SelectFilter(id=NULL, label=NULL, value=NULL, startDate=NULL, type=NULL,
endDate=NULL, classes=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}

\item{startDate}{Logical | numeric | character | named list | unnamed list. The startDate displayed in the input.}

\item{type}{Logical | numeric | character | named list | unnamed list. The type of dateformat in the input.}

\item{endDate}{Logical | numeric | character | named list | unnamed list. The endDate displayed in the input.}

\item{classes}{Named list. The classes displayed in the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

